Middleware creados, registrados en el kernel, implementados en las rutas y probados

+++ routes/api.php
@@ -22,13 +22,13 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
 Route::get('/usuarios/{id?}', 'API\UsuarioController@index')->where(['id','[0-9]+']);
 Route::post('/usuarios', 'API\UsuarioController@crear');
 Route::put('/usuarios/{id}', 'API\UsuarioController@editar')->where(['id','[0-9]+']);
-Route::delete('/usuarios/{id}', 'API\UsuarioController@eliminar')->where(['id','[0-9]+']);
+Route::delete('/usuarios/{id}', 'API\UsuarioController@eliminar')->where(['id','[0-9]+'])->middleware('verifica.admin');
 
 //CRUD PRODUCTOS
 Route::get('/productos/{id?}', 'API\ProductoController@index')->where(['id','[0-9]+']);
 Route::post('/productos', 'API\ProductoController@crear');
 Route::put('/productos/{id}', 'API\ProductoController@editar')->where(['id','[0-9]+']);
-Route::delete('/productos/{id}', 'API\ProductoController@eliminar')->where(['id','[0-9]+']);
+Route::delete('/productos/{id}', 'API\ProductoController@eliminar')->where(['id','[0-9]+'])->middleware('verifica.cantidad');
 
 //CRUD COMENTARIOS
 Route::get('/comentarios/{id?}', 'API\ComentarioController@index')->where(['id','[0-9]+']);
